Broken savefile bug fix
Fixed a stupid bug where the patch version number was being written/read as a string instead of a number. This is apparently fine if the number is zero, but if the number is 1+ trying to open a savefile exits without a useful message. FrogComposband uses a flavor for the patch version but Oposband just uses a number and I had missed updating the field type in two locations.
Also added a "Precompiled" folder with the compiled windows zips.