Table balancing
Removed / commented code for projectile blocking effect of tables until its feasible to apply to magical projectiles as well. Also 60% was OP, when it comes back it will probably be half that. Melee blocking effect of tables also removed until projectiles are added again, and will likewise be nerfed to half its effectiveness. 
Jumping onto a table is now a 1d30 vs DEX check, no need for a separate skill for this.