< and > now the same command
Merged the practically identical < and > commands. Now you can use either one to take a staircase, or to toggle between normal and world map view on the surface. 
Also updated help files on the changes to the 'a' command and the swapping of Q and ^Q (you check quest status way more often than you quit).